HSEARCH-3614 Test annotation reading with every reflection strategy
