+++ mapper/orm/src/test/java/org/hibernate/search/mapper/orm/model/impl/HibernateOrmBootstrapIntrospectorAccessTypeTest.java
@@ -8,9 +8,6 @@ package org.hibernate.search.mapper.orm.model.impl;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import javax.persistence.Access;
@@ -26,52 +23,14 @@ import javax.persistence.MapKeyClass;
 import javax.persistence.OneToMany;
 import javax.persistence.Transient;
 
-import org.hibernate.annotations.common.reflection.ReflectionManager;
-import org.hibernate.boot.Metadata;
-import org.hibernate.boot.MetadataSources;
-import org.hibernate.boot.registry.StandardServiceRegistry;
-import org.hibernate.boot.registry.StandardServiceRegistryBuilder;
-import org.hibernate.boot.spi.MetadataImplementor;
-import org.hibernate.cfg.AvailableSettings;
-import org.hibernate.dialect.H2Dialect;
-import org.hibernate.search.engine.cfg.ConfigurationPropertySource;
-import org.hibernate.search.mapper.orm.cfg.spi.HibernateOrmMapperSpiSettings;
 import org.hibernate.search.mapper.pojo.model.spi.PojoRawTypeModel;
 import org.hibernate.search.util.common.reflect.spi.ValueReadHandle;
-import org.hibernate.search.util.common.impl.Closer;
 
-import org.junit.After;
 import org.junit.Assert;
 import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.junit.runners.Parameterized;
 
-@RunWith(Parameterized.class)
-public class HibernateOrmBootstrapIntrospectorAccessTypeTest {
-
-	@Parameterized.Parameters(name = "Reflection strategy = {0}")
-	public static List<Object[]> data() {
-		return Arrays.asList( new Object[][] {
-				{ null },
-				{ "method-handle" },
-				{ "java-lang-reflect" }
-		} );
-	}
-
-	private final List<AutoCloseable> toClose = new ArrayList<>();
-
-	private final String reflectionStrategyName;
-
-	public HibernateOrmBootstrapIntrospectorAccessTypeTest(String reflectionStrategyName) {
-		this.reflectionStrategyName = reflectionStrategyName;
-	}
-
-	@After
-	public void cleanup() throws Exception {
-		try ( Closer<Exception> closer = new Closer<>() ) {
-			closer.pushAll( AutoCloseable::close, toClose );
-		}
-	}
+public class HibernateOrmBootstrapIntrospectorAccessTypeTest
+		extends AbstractHibernateOrmBootstrapIntrospectorPerReflectionStrategyTest {
 
 	@Test
 	public void entity_defaultFieldAccess() {
@@ -267,37 +226,6 @@ public class HibernateOrmBootstrapIntrospectorAccessTypeTest {
 		assertThat( valueReadHandle.get( nestedEmbeddable ) ).isEqualTo( nestedEmbeddable.getPropertyWithDefaultMethodAccess() );
 		valueReadHandle = nestedEmbeddableTypeModel.getProperty( "propertyWithNonDefaultFieldAccess" ).getHandle();
 		assertThat( valueReadHandle.get( nestedEmbeddable ) ).isEqualTo( nestedEmbeddable.propertyWithNonDefaultFieldAccess );
-	}
-
-	private HibernateOrmBootstrapIntrospector createIntrospector(Class<?> ... entityClasses) {
-		StandardServiceRegistryBuilder registryBuilder = new StandardServiceRegistryBuilder();
-		// Some properties that are not relevant to our test, but necessary to create the Metadata
-		registryBuilder.applySetting( AvailableSettings.DIALECT, H2Dialect.class );
-		StandardServiceRegistry serviceRegistry = registryBuilder.build();
-		toClose.add( serviceRegistry );
-
-		MetadataSources metadataSources = new MetadataSources( serviceRegistry );
-		for ( Class<?> entityClass : entityClasses ) {
-			metadataSources.addAnnotatedClass( entityClass );
-		}
-		Metadata metadata = metadataSources.buildMetadata();
-
-		MetadataImplementor metadataImplementor = (MetadataImplementor) metadata;
-		ReflectionManager reflectionManager = metadataImplementor.getTypeConfiguration()
-				.getMetadataBuildingContext().getBootstrapContext().getReflectionManager();
-
-		Map<String, Object> properties = new HashMap<>();
-		if ( reflectionStrategyName != null ) {
-			properties.put(
-					HibernateOrmMapperSpiSettings.Radicals.REFLECTION_STRATEGY,
-					reflectionStrategyName
-			);
-		}
-
-		return HibernateOrmBootstrapIntrospector.create(
-				metadata, reflectionManager,
-				ConfigurationPropertySource.fromMap( properties )
-		);
 	}
 
 	private static <T> T methodShouldNotBeCalled() {
